three classes added
