+++ .gitignore
@@ -1,0 +1,31 @@
+## JAVA
+.mtj.tmp 
+*.class 
+*.jar 
+*.war 
+*.ear 
+*.nar 
+hs_err_pid* 
+
+## MAVEN
+target/ 
+pom.xml.tag 
+pom.xml.releaseBackup 
+pom.xml.versionsBackup 
+pom.xml.next 
+pom.xml.bak 
+release.properties 
+dependency-reduced-pom.xml 
+buildNumber.properties 
+.mvn/timing.properties 
+.mvn/wrapper/maven-wrapper.jar 
+
+## INTELLIJ
+out/ 
+.idea/ 
+.idea_modules/ 
+*.iml 
+*.ipr 
+*.iws 
+
+
