+++ README.md
@@ -1,1 +1,5 @@
-# stormIT-homeworks
+StormIT Homeworks Repo.
+My own ideas and solutions.
+
+-- Contents: --
+#Matrix
