+++ src/main/java/arrays/UserInput.java
@@ -6,6 +6,9 @@ import java.util.Scanner;
 public class UserInput {
 
     private Scanner scan = new Scanner(System.in);
+    private static final String ROWS_NUMBER_MSG = "Podaj liczbę wierszy (2 - 10)";
+    private static final String COLUMNS_NUMBER_MSG = "Podaj liczbę kolumn (2 - 10)";
+    private static final String GIVE_ME_INTEGER_MSG = "Podaj liczbę całkowitą z zakresu 2 -10!";
 
     private int readInteger(String msg) {
         int dim = 0;
@@ -19,11 +22,11 @@ public class UserInput {
                     correctInput = true;
                 }
                 else {
-                    System.out.println("Podaj liczbę całkowitą z zakresu 2 -10!");
+                    System.out.println(GIVE_ME_INTEGER_MSG);
                     continue;
                 }
             } catch (InputMismatchException ex) {
-                System.out.println("Podaj liczbę całkowitą z zakresu 2 -10!");
+                System.out.println(GIVE_ME_INTEGER_MSG);
             } finally {
                 scan.nextLine();
             }
@@ -32,8 +35,8 @@ public class UserInput {
     }
 
     double[][] giveDimension() {
-        int dim1 = readInteger("Podaj liczbę kolumn (2 - 10)");
-        int dim2 = readInteger("Podaj liczbę wierszy (2 - 10)");
+        int dim1 = readInteger(COLUMNS_NUMBER_MSG);
+        int dim2 = readInteger(ROWS_NUMBER_MSG);
         return new double[dim1][dim2];
     }
 }
